Removed unnecessary files
