--- graphene/utils/is_base_type.py
@@ -1,3 +1,0 @@
-
-def is_base_type(bases, _type):
-    return any(b for b in bases if isinstance(b, _type))
